pyproject.toml
[tool.poetry]
name = "zappts"
version = "0.1.0"
description = ""
authors = ["higorvinicius <higor.tecinfor@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask-RESTful = "^0.3.9"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.0.0"
Flask-Cors = "^3.0.10"
dynaconf = "^3.1.4"
psycopg2-binary = "^2.8.6"
Flask-JWT-Extended = "^4.2.1"
Flask-Admin = "^1.5.8"
python-slugify = "^5.0.2"
Flask-Mail = "^0.9.1"
black = "^21.5b1"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.17.1"
pytest = "^6.2.4"
pytest-flask = "^1.2.0"
pytest-regressions = "^2.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
